Springboot Day 25

+++ src/main/java/com/example/pcmspringbot1/repo/GroupMenuRepo.java
@@ -3,7 +3,7 @@ package com.example.pcmspringbot1.repo;
 import com.example.pcmspringbot1.model.GroupMenu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GroupMenuRepo extends JpaRepository<GroupMenu, Integer> {
+public interface GroupMenuRepo extends JpaRepository<GroupMenu, Long> {
 
 
 }
